Update jquery.table2excel.js
Extended "exclude" capabilities down to the <td> level.  Original functionality only allowed you to "exclude" an entire <tr>.
Now if any item in a <td> has the exclude class on it, that <td> will be exported EMPTY.

This is very useful if you habe buttons or other controls in your table and do not want to export them.